Two conflicts, both from the kernel worker building on this file.

`src/metaldevices.rs` is now what both profiles share and nothing more: the
refusal codes `userland/metalprobe` exits with, the device inventory, the
shutdown's own account, and `Bound` — which `src/metalkernel.rs` judges against
too, and which is here so that one arithmetic answers for what a ceiling means.
What went is the device suite's own `Job`/`JOBS`/`judge`: every device number is
priced by name in `tests/metal-profile.toml` now, and a second profile in a
second grammar was the thing that file exists to end. `Refused::of` is public
because the harness's judge reads an exit code with it.

`tests/metal-profile.toml` takes both blocks: the kernel worker's latency and
TLB rows and this branch's device spans.

+++ tests/common/devices.rs
@@ -71,7 +71,7 @@ pub fn on_metal(back: &metal::Readback) -> Result<(), String> {
 
 /// One job's exit code as the number it is, or why it is not one.
 fn measurement(job: &str, code: i32) -> Result<u64, String> {
-    if let Some(refused) = metaldevices::Refused::of(code) {
+    if let Some(refused) = metaldevices::Refused::of(i64::from(code)) {
         return Err(format!("{job}: the job refused — {refused}"));
     }
     u64::try_from(code)
